requirements.txt
flask
requests
playsound==1.2.2
clipboard
screeninfo
configparser
mysql-connector
